Crate ra_cfg

ra_cfg defines conditional compiling options, cfg attibute parser and evaluator

Structs

CfgOptions
Configuration options used for conditional compilition on items with cfg attributes. We have two kind of options in different namespaces: atomic options like unix, and key-value options like target_arch="x86".

Enums

CfgExpr

Functions

parse_cfg